refactor: refactored projects to inside portfolio directory and renamed tasks and issues directory
